+++ packages/api/lambdas/sf-event-sqs-to-db-records/write-granules.js
@@ -2,7 +2,6 @@
 
 const AggregateError = require('aggregate-error');
 
-const { describeExecution } = require('@cumulus/aws-client/StepFunctions');
 const log = require('@cumulus/common/log');
 const {
   tableNames,
