+++ packages/api/tests/lambdas/sf-event-sqs-to-db-records/test-utils.js
@@ -7,8 +7,6 @@ const uuidv4 = require('uuid/v4');
 const {
   localStackConnectionEnv,
   getKnexClient,
-  tableNames,
-  doesRecordExist,
 } = require('@cumulus/db');
 
 const { migrationDir } = require('../../../../../lambdas/db-migration');
